refactor: mover componentes StickyBar, popup y videoPlayer a la carpeta ui
