+++ src/templates/airportAndOffice.jsx
@@ -5,7 +5,7 @@ import ContentToggle from "../components/ui/ContentToggle"
 import ContactElements from "../components/ui/ContactElements"
 import AddressBlock from "../components/ui/AddressBlock"
 import GoogleMapBlock from "../components/ui/GoogleMapBlock"
-import StickyBar from "../components/StickyBar"
+import StickyBar from "../components/ui/StickyBar"
 
 const AirportAndOffice = ({ pageContext }) => {
   const { data, statusElement } = useLocalizedQuery(
